backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
openai==1.59.7
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
email-validator==2.2.0
